Add: a new alias for browse-kill-ring.
I was previously using the ido business to type b-k-r, but I use it
frequently enough that it really makes sense to just use an
alias. Also, I'm too lazy to find a new keyboard shortcut, and this is
much closer to my current workflow, so I think it's more useful to do
it this way.