Add: redefine M-x term not to ask for program command; breaks windows
I'm tired of M-x term always asking me for the program command to run,
as it's almost always /bin/bash, so I've gone in and redefined term so
that it won't ask. However, this makes it much less likely to work on
a Windows box, as they won't have bash there. I'll fix that if and
when I need to. For now, this works on OS X and probably linux.