Fix hippie expand try functions list
Previously, the first entry was yas/hippie-expand something or other,
but I'm still not really familiar with that business enough for that
to have worked. More importantly, I don't have yasnippet installed, so
that probably would not have gone very well at all for me. Now, it
tries to do expansions without yasnippet.