Add yasnippet to el-get-sources.
Picked up yasnippet from https://github.com/capitaomorte/yasnippet, as
the version in el-get packages is outdated and pulls from
code.google.com, where it apparently no longer is. Note that the code
that he suggests doesn't work (due to error in type field? or website
field? one or the other...), so I changed it a bit, and now it does
work.