Introduce compile-again to replace recompile
Recompile calls the compile-command based on what buffer you're
currently in, reverting back to the normal compile behavior which
defaults to make -k. Using smart-compile, the normal compile behavior
is almost certainly wrong, so compile-again just repeats the most
recently called compile command, which is the functionality that I
want.