Stop requiring things and eval-after-load them first.
Of arguable efficacy, this is supposed to speed up the loading of the
init.el file. The only problem is that a lot of the things that I put
in customize.el were from all over, so it'll be a lot harder to find
the things I'm looking for. To that end, I'm putting in comments that
describe where things went.