Fix issue with M-k sometimes complaining
Was getting "Wrong type argument: frame-live-p, #<dead frame Emacs:"
or similar errors, sometimes it was buffer-live-p and a dead
buffer. Not sure that this is the best solution - in the
kill-this-buffer defun, it's doing an if on
menu-bar-non-minibuffer-window-p and when it comes back with a dead
frame the if apparently doesn't like it. Well, this variable seems to
solve the issue, at least for the time being.