Compensate for discover/dired-x clashes
The error was "Key sequence * O starts with non-prefix key *". The
problem was that loading discover first bound * to
makey-key-mode-popup-dired-marking.

    ("*" "mark ..." makey-key-mode-popup-dired-marking)

and

    (dired-marking dired-mode dired-mode-hook "*")
    (dired-regexp dired-mode dired-mode-hook "%")

Since we didn't explicitly load dired-x first, magit or any other thing
that tried to use dired-jump would attempt to bind to "*O"

    (define-key dired-mode-map "*O" 'dired-mark-omitted)

and we'd run into a problem because the "*" was already specifically
bound to something. The solution is to explicitly load dired-x ourselves
and let it bind its keybindings. Afterwards, we can overwrite things as
we please in discover mode.